Moved execution test up to AbstractReportingMojo to skip overall reporting too
